+++ frontend/src/googleAuth/initAuth.ts
@@ -1,9 +1,10 @@
 import JsonToken from "../../types/JsonTokenType";
+import { handlePostRequest } from "../main";
 
 const client_id = import.meta.env.VITE_GOOGLE_AUTH_CLIENT_ID;
 
 
-const saveToLocalStorage = (decoded : JsonToken)=>{
+const saveToLocalStorage = (decoded : JsonToken) => {
 	const email = decoded.email;
 	const familyName = decoded.family_name;
 	const givenName = decoded.given_name;
@@ -15,35 +16,15 @@ const saveToLocalStorage = (decoded : JsonToken)=>{
 	localStorage.setItem("imgProfil", imgProfil);
 }
 
-const decodeJWTToken = (token : string) => {
-	const base64Url = token.split('.')[1];
-	const base64 = base64Url.replace(/-/g, '+').replace(/_/g, '/');
-	const jsonPayload = decodeURIComponent(
-		atob(base64)
-			.split('')
-			.map((c) => {
-				return '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2);
-			})
-			.join('')
-	);
-
-	const decoded = JSON.parse(jsonPayload);
-	saveToLocalStorage(decoded);
-
-	return decoded.email;
-}
-
-
 function handleCredentialResponse(response: any) {
 	if(response.credential)
 	{
-		decodeJWTToken(response.credential);
 		const popup = document.getElementById("loginPopup");
-		if (popup) {
+		const token = response.credential;
+		handlePostRequest("/api/auth/login", token);
+		if (popup) 
 			popup.style.display = "none";
-		}
 	}
-	
 }
 
 export function loadGoogleSignInScript(): Promise<void> {
@@ -101,8 +82,6 @@ export function initGoogleAuth() {
 		popup.style.top = "50%";
 		popup.style.left = "50%"; 
 		popup.style.transform = "translateX(-50%)"; 
-    
-	
 		google.accounts.id.renderButton(
 			document.getElementById("loginPopup")!,
 			{ theme: "filled_black", size: "large" }
@@ -128,5 +107,4 @@ export function setupLogoutButton() {
 		localStorage.removeItem("givenName");
 		localStorage.removeItem("imgProfil");
 	}
-		//localStorage.removeItem("i18nextLng"); //! Uncomment if you want to remove the language from local storage
 }
