+++ services/auth/src/server.ts
@@ -12,10 +12,10 @@ export default async function startServer() {
 	});
 
 	fastify.register(loginRoute);
-	fastify.register(doesuserexistRoute);
 
 	try {
 		await fastify.listen({ port: 3000, host: '0.0.0.0' });
+		fastify.ready().then(() => console.log(fastify.printRoutes()));
 		fastify.log.info('auth service is running on port 3000');
 	} catch (err) {
 		fastify.log.error(err);
